requirements.txt
matplotlib==3.7.1
numpy==1.24.2
pandas==1.5.3
Pillow==9.5.0
scikit_learn==1.2.2
statsmodels==0.14.0
streamlit==1.25.0
openpyxl==3.1.2
lightgbm==3.3.5
xgboost==2.0.3
plotly==5.14.1
betacal==1.1.0
shap==0.44.1
